Add primitive sticker adding functionality
User can now ask the bot to add a sticker to their personal set. The
user must first send a photo and afterwards the corresponding emoji.
Then the bot will create a sticker from that information and add it to
the personal set of the user named after _sticker_set_name.