Bump version to 0.1.0
Probably the first release ready version. Not sure if it is in any way
useful to keep track of the version number since it's unlikely that this
will be currently released as a normal Python package as was initially
planned. However, for VCS purposes it doesn't hurt and is at least
clearer to follow compared to the Docker tags.