Add init routine to fokit
The persistence of user_data and chat_data cannot handle arbitrary data
types as they currently have to pickle the data. Therefore, as a
workaround we need a separate init functionality in fokit to recreate
the Jobs in the JobQueue during startup as they are lost in shutdowns.

Contains also minor naming changes.