Add a crawler and indexing to fokit
Now we can crawl all of the available comics on the host site, store and
index them locally for later use. Therefore, in order to post the latest
comic we would only need to update the index and then post the latest
available image pointed there.

This indexing method allows us to easily post random comics as well,
which greatly affected this design choice. Also by storing the data
locally we avoid unnecessary delays when a request comes from a user.