Add timezone configuration in Dockerfile
The container was running using UTC time previously. This caused issues
as the daily Fok-It posting was assuming local time being used.
Furthermore, even if the posting time would be configured to be UTC then
the DST would cause issues twice a year. Thus, using local time is
probably the best solution.